Split easybib config into to classes
one for overall config logic, one for file generation

also includes rubocop fixes
