+++ .rubocop_todo.yml
@@ -7,14 +7,3 @@ Style/AccessorMethodName:
     - 'wt-data/libraries/appstodeploy.rb'
     - 'wt-data/tests/test_accessor.rb'
     - 'wt-data/tests/test_dataobject.rb'
-
-# Offense count: 5
-# Configuration parameters: EnforcedStyle, SupportedStyles.
-# SupportedStyles: module_function, extend_self
-Style/ModuleFunction:
-  Exclude:
-    - 'easybib/libraries/config.rb'
-    - 'easybib/libraries/easybib.rb'
-    - 'easybib/libraries/ppa.rb'
-    - 'easybib/libraries/sns.rb'
-    - 'easybib/libraries/upstream.rb'
